requirements.txt
google-cloud-dialogflow==2.15.1
python-telegram-bot==13.13
vk-api==11.9.9